# openapi.yaml
openapi: 3.0.3
info:
  title: Taskryse API
  version: 1.0.0
  description: >
    Backend contract for Taskryse. Money in minor units (kobo/cents).
    Core rule: plans grant ELIGIBILITY only; earnings come from APPROVED work
    against FUNDED sponsor escrow; promo credits are non-withdrawable.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    Task:
      type: object
      properties:
        id: {type: string}
        title: {type: string}
        categoryId: {type: string}
        difficulty: {type: string, enum: [BEGINNER, INTERMEDIATE, EXPERT]}
        payMinor: {type: integer}
        currency: {type: string}
        deadline: {type: string, format: date-time}
        totalSlots: {type: integer}
        slotsTaken: {type: integer}
        requiresQualification: {type: boolean}
        eligibility: {type: string, enum: [ELIGIBLE, QUALIFICATION_REQUIRED, PLAN_LOCKED], description: computed per caller}
...
paths:
...
  /subscriptions/{id}/upgrade: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: ok } } } }
  /subscriptions/{id}/cancel: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: ok } } } }
...
  /tasks:
    get:
      summary: Marketplace. Returns per-caller eligibility.
      parameters:
        - {name: category, in: query, schema: {type: string}}
        - {name: difficulty, in: query, schema: {type: string}}
        - {name: minPay, in: query, schema: {type: integer}}
        - {name: country, in: query, schema: {type: string}}
        - {name: qualification, in: query, schema: {type: boolean}}
      responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Task' } } } } } }
  /tasks/{id}: { get: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } } } } }
  /tasks/{id}/accept: { post: { summary: Accept task (checks eligibility, slots, plan), parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '201': { description: assignment created }, '409': { description: not eligible / slots full } } } }
  /qualifications/{id}/attempt: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: scored } } } }
...